重写员工同步

+++ linkwe-service/src/main/java/com/linkwechat/domain/system/dept/query/SysDeptQuery.java
@@ -16,5 +16,5 @@ import java.util.List;
 public class SysDeptQuery {
 
     @ApiModelProperty("部门ID")
-    private List<Integer> deptIds;
+    private List<Long> deptIds;
 }
